requirements.txt
cryptography>=41.0.0
pyyaml>=6.0 